- moved `LivePreviewSection` to be a UI component only 	- this will allow us to reuse it in multiple places
